requirements.txt
celery==3.1.26.post2
coverage==5.5
Django==3.0.5
django-celery==3.3.1
django-rest-passwordreset==1.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
Pillow==8.4.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytest==6.2.3
pytest-cov==2.11.1
pytest-django==4.1.0
python-decouple==3.4
